third_party/blink/web_tests/platform/generic/http/tests/security/xss-DENIED-window-index-assign-expected.txt

ALERT: undefined
ALERT: PASS: Access to parent[0].f() threw exception.

This test passes if the access is forbidden.
